backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.11
web3==6.0.0
eth-account==0.8.0
py-algorand-sdk==2.1.0
python-multipart==0.0.6
# Lute wallet SDK (placeholder) - replace with the real package name when known
# Example: lute-sdk==0.1.0
